+++ src/maggma/api/resource/post_resource.py
@@ -106,7 +106,7 @@ class PostOnlyResource(Resource):
             operator_meta = {}
 
             for operator in self.query_operators:
-                data = operator.post_process(data)
+                data = operator.post_process(data, query)
                 operator_meta.update(operator.meta())
 
             meta = Meta(total_doc=count)