+++ src/maggma/stores/mongolike.py
@@ -5,15 +5,15 @@ Stores are a default access pattern to data and provide
 various utilities
 """
 
-import json
 from pathlib import Path
-
 import yaml
 from itertools import chain, groupby
 from socket import socket
+import warnings
 from typing import Any, Dict, Iterator, List, Optional, Tuple, Union
 
 import mongomock
+import orjson
 from monty.dev import requires
 from monty.io import zopen
 from monty.json import MSONable, jsanitize
@@ -638,9 +638,15 @@ class MemoryStore(MongoStore):
             generator returning tuples of (key, list of elemnts)
         """
         keys = keys if isinstance(keys, list) else [keys]
+
+        if properties is None:
+            properties = []
+        if isinstance(properties, dict):
+            properties = list(properties.keys())
+
         data = [
             doc
-            for doc in self.query(properties=keys, criteria=criteria)
+            for doc in self.query(properties=keys + properties, criteria=criteria)
             if all(has(doc, k) for k in keys)
         ]
 
@@ -670,28 +676,60 @@ class JSONStore(MemoryStore):
     A Store for access to a single or multiple JSON files
     """
 
-    def __init__(self, paths: Union[str, List[str]], file_writable=False, **kwargs):
+    def __init__(
+        self,
+        paths: Union[str, List[str]],
+        read_only: bool = True,
+        **kwargs,
+    ):
         """
         Args:
             paths: paths for json files to turn into a Store
-            file_writable: whether this JSONStore is "file-writable". When a JSONStore
-                is "file-writable", the json file will be automatically updated
-                everytime a write-like operation is performed. Note that only
-                JSONStore with a single JSON file is compatible with file_writable
-                True. Note also that when file_writable is False, the JSONStore
-                can still apply MongoDB-like writable operations (e.g. an update)
-                as it behaves like a MemoryStore, but it will not write those changes
-                to the file.
+            read_only: whether this JSONStore is read only. When read_only=True,
+                       the JSONStore can still apply MongoDB-like writable operations
+                       (e.g. an update) because it behaves like a MemoryStore,
+                       but it will not write those changes to the file. On the other hand,
+                       if read_only=False (i.e., it is writeable), the JSON file
+                       will be automatically updated every time a write-like operation is
+                       performed.
+
+                       Note that when read_only=False, JSONStore only supports a single JSON
+                       file. If the file does not exist, it will be automatically created
+                       when the JSONStore is initialized.
         """
         paths = paths if isinstance(paths, (list, tuple)) else [paths]
         self.paths = paths
-        if file_writable and len(paths) > 1:
+
+        # file_writable overrides read_only for compatibility reasons
+        if "file_writable" in kwargs:
+            file_writable = kwargs.pop("file_writable")
+            warnings.warn(
+                "file_writable is deprecated; use read only instead.",
+                DeprecationWarning,
+            )
+            self.read_only = not file_writable
+            if self.read_only != read_only:
+                warnings.warn(
+                    f"Received conflicting keyword arguments file_writable={file_writable}"
+                    f" and read_only={read_only}. Setting read_only={file_writable}.",
+                    UserWarning,
+                )
+        else:
+            self.read_only = read_only
+        self.kwargs = kwargs
+
+        if not self.read_only and len(paths) > 1:
             raise RuntimeError(
                 "Cannot instantiate file-writable JSONStore with multiple JSON files."
             )
-        self.file_writable = file_writable
-        self.kwargs = kwargs
-        super().__init__(collection_name="collection", **kwargs)
+
+        # create the .json file if it does not exist
+        if not self.read_only and not Path(self.paths[0]).exists():
+            with zopen(self.paths[0], "w") as f:
+                data: List[dict] = []
+                bytesdata = orjson.dumps(data)
+                f.write(bytesdata.decode("utf-8"))
+        super().__init__(**kwargs)
 
     def connect(self, force_reset=False):
         """
@@ -702,9 +740,19 @@ class JSONStore(MemoryStore):
             with zopen(path) as f:
                 data = f.read()
                 data = data.decode() if isinstance(data, bytes) else data
-                objects = json.loads(data)
+                objects = orjson.loads(data)
                 objects = [objects] if not isinstance(objects, list) else objects
-                self.update(objects)
+                try:
+                    self.update(objects)
+                except KeyError:
+                    raise KeyError(
+                        f"""
+                        Key field '{self.key}' not found in {f.name}. This
+                        could mean that this JSONStore was initially created with a different key field.
+                        The keys found in the .json file are {list(objects[0].keys())}. Try
+                        re-initializing your JSONStore using one of these as the key arguments.
+                        """
+                    )
 
     def update(self, docs: Union[List[Dict], Dict], key: Union[List, str, None] = None):
         """
@@ -720,7 +768,7 @@ class JSONStore(MemoryStore):
                  field is to be used
         """
         super().update(docs=docs, key=key)
-        if self.file_writable:
+        if not self.read_only:
             self.update_json_file()
 
     def remove_docs(self, criteria: Dict):
@@ -733,7 +781,7 @@ class JSONStore(MemoryStore):
             criteria: query dictionary to match
         """
         super().remove_docs(criteria=criteria)
-        if self.file_writable:
+        if not self.read_only:
             self.update_json_file()
 
     def update_json_file(self):
@@ -744,7 +792,8 @@ class JSONStore(MemoryStore):
             data = [d for d in self.query()]
             for d in data:
                 d.pop("_id")
-            json.dump(data, f)
+            bytesdata = orjson.dumps(data)
+            f.write(bytesdata.decode("utf-8"))
 
     def __hash__(self):
         return hash((*self.paths, self.last_updated_field))